test: keep assessor review regression coverage bounded
