+++ internal/service/memoryservice/submission_assessment_worker_test.go
@@ -51,29 +51,6 @@ func TestSubmissionAssessmentWorkerAssessesWholeRunAndCommitsAtomically(t *testi
 	assert.Len(t, catalog.predicateOptionInputs, 1)
 }
 
-func TestSubmissionAssessmentWorkerSharesOneEntityResolutionForRepeatedGrounding(t *testing.T) {
-	_, assessments, _, provider, worker := submissionAssessmentWorkerFixture(t)
-	provider.response = func(req verifier.SemanticAssessmentRequest) (verifier.SemanticAssessmentResponse, error) {
-		response := submissionAssessmentValidResponse(req, false)
-		for index, entity := range req.SubmittedEntities {
-			if entity.Name != "Alpha" {
-				continue
-			}
-			groundingRef := entity.Groundings[0].GroundingRef
-			response.EntityResults[index].GroundingRef = &groundingRef
-		}
-		return response, nil
-	}
-
-	processed, err := worker.ProcessNextSubmissionAssessmentPlacement(context.Background())
-
-	require.NoError(t, err)
-	require.True(t, processed)
-	require.Len(t, assessments.commits, 1)
-	assert.Len(t, assessments.commits[0].EntityResolutions, 5)
-	assert.Empty(t, assessments.completions)
-}
-
 func TestSubmissionAssessmentWorkerTerminalizesPredicateOptionOverflowBeforeProvider(t *testing.T) {
 	_, assessments, catalog, provider, worker := submissionAssessmentWorkerFixture(t)
 	catalog.predicateComplete = false
